feature/#186/opps-comments-section: Move common ui to EntityCommenty
