+++ src/components/Dashboard/Profile/sections/Comments/common/index.ts
@@ -2,6 +2,7 @@ export * from "./styles";
 export { Comment } from "./Comment";
 export { CommentActionMenu } from "./CommentActionMenu";
 export { DeleteCommentDialog } from "./DeleteCommentDialog";
+export { EntityComments } from "./EntityComments";
 export { useCommentDelete } from "./hooks/useCommentDelete";
 export { useCommentEdit } from "./hooks/useCommentEdit";
 export { useCommentMenu } from "./hooks/useCommentMenu";
